requirements.txt
annotated-types==0.6.0
anyio~=4.3.0
asgiref==3.7.2
boto3==1.34.59
botocore==1.34.59
certifi~=2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.10
docstring-to-markdown==0.15
exceptiongroup==1.2.0
h11==0.14.0
httpcore~=0.16.3
httpx~=0.23.3
idna==3.6
importlib_metadata==7.0.2
jedi==0.19.1
jmespath==1.0.1
myenv==201902.7
nodeenv==1.8.0
openai~=1.25.2
parso==0.8.3
pluggy==1.4.0
pydantic~=1.10.15
pydantic_core==2.16.2
pyright==1.1.352
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio~=1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
ujson==5.9.0
urllib3~=1.26.14
zipp==3.17.0
channels~=4.1.0
matplotlib~=3.8.4
numpy~=1.26.4
streamlit~=1.25.0
sympy~=1.12
aioconsole~=0.7.0
panel~=1.3.8
djangorestframework~=3.15.1
pillow~=10.2.0
pip~=24.0
cryptography~=42.0.5
lxml~=4.9.4
keyring~=25.1.0
setuptools~=65.5.0
nh3~=0.2.17
attrs~=23.2.0
docutils~=0.19
tornado~=6.4
Jinja2~=3.1.3
filelock~=3.13.3
redis~=5.0.3
rich~=13.7.1
pygments~=2.17.2
ipython~=8.12.3
PyYAML~=6.0.1
FLAML~=2.1.2
scipy~=1.12.0
scikit-learn~=1.4.1.post1
pandas~=2.2.1
diskcache~=5.6.3
termcolor~=2.4.0
docker~=7.0.0
tiktoken~=0.5.2
click~=8.1.7
mdurl~=0.1.2
pytz~=2024.1
overrides~=7.7.0
cffi~=1.16.0
regex~=2023.12.25
twine~=5.0.0
rfc3986~=1.5.0
pkginfo~=1.10.0
build~=1.2.1
packaging~=23.2
websockets~=12.0
python-dotenv~=1.0.0
websocket-client~=1.7.0
coding~=0.3
docopt~=0.6.2
autopep8~=2.1.0
beautifulsoup4~=4.12.3
decorator~=5.1.1
Deprecated~=1.2.14
pyparsing~=3.1.2
pycodestyle~=2.11.1
python-multipart~=0.0.9
watchdog~=4.0.0
chromadb~=0.4.3
plotly~=5.20.0
bs4~=0.0.1
langchain~=0.1.17
pypdf2~=3.0.1
emoji~=2.11.1
transformers~=4.39.2
psutil~=5.9.8
fastembed~=0.3.4
pypdf~=4.1.0
unstructured~=0.13.2
h2~=4.1.0
validators~=0.28.1
chardet~=5.2.0